+++ CHANGELOG.md
@@ -3,3 +3,13 @@
 ## 0.0.1
 
 - Initial release
+
+## 0.0.5
+
+- new tree view btns to
+    + open settings file
+    + close all opened editors and reset layout
+    + setup file group position, set "`workbench.editor.closeEmptyGroups: false`"
+- new configs
+    + hideSideBarAfterOpen
+    + layingOutType